Make container cleanup optional

+++ run_in_docker.sh
@@ -2,6 +2,9 @@
 
 # This script assumes that a Go service is already built, and the Python service can be interpreted without errors
 
+# This parameter determines whether to shut down containers when the script finishes
+cleanup_enabled=${CLEANUP_ENABLED:-1}
+
 # Function to get docker compose profile to add based on service name specified by the user
 with_profile() {
   local target="$1"
@@ -127,7 +130,9 @@ copy_files() {
 }
 
 cleanup() {
-  docker compose down
+  if [[ $cleanup_enabled != 0 ]]; then
+    docker compose down
+  fi
 }
 
 # Step 1: Specify the make target for tests to run
